Rename resolution_center.py to issue_handler.py
