+++ homeassistant/components/resolution_center/__init__.py
@@ -6,8 +6,8 @@ from homeassistant.helpers.typing import ConfigType
 
 from . import websocket_api
 from .const import DOMAIN
+from .issue_handler import async_create_issue, async_delete_issue  # noqa: F401
 from .issue_registry import async_load as async_load_issue_registry
-from .resolution_center import async_create_issue, async_delete_issue  # noqa: F401
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
