+++ tests/components/resolution_center/test_init.py
@@ -6,9 +6,7 @@ from homeassistant.components.resolution_center import (
     async_delete_issue,
 )
 from homeassistant.components.resolution_center.const import DOMAIN
-from homeassistant.components.resolution_center.resolution_center import (
-    async_dismiss_issue,
-)
+from homeassistant.components.resolution_center.issue_handler import async_dismiss_issue
 from homeassistant.const import MAJOR_VERSION, MINOR_VERSION, PATCH_VERSION
 from homeassistant.core import HomeAssistant
 from homeassistant.setup import async_setup_component
